components/requirements.txt
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-celery
celery[redis]
redis
numba
numpy
mysql-connector-python
ccxt
pyyaml
pytest
eth-account
hyperliquid-python-sdk